v1/requirements.txt
eth-hash>=0.7.0

ccxt>=4.4.35
coinbase-advanced-py>=1.8.2
lxml>=5.3.0
pytest>=8.3.4
pytest-mock>=3.14.0 
python-dateutil>=2.9.0
pytz>=2024.2
requests>=2.32.3
web3>=7.6.0
pyjwt>=2.10.1
cryptography>=44.0.0
